refactor(functions): split AgentService — API into @dxos/functions, impl stays in @dxos/assistant
- Move FunctionNotFoundError to @dxos/operation to break the circular dep @dxos/blueprints → @dxos/functions
- Add @dxos/blueprints as a dep of @dxos/functions (now safe without the circular dep)
- Create @dxos/functions/src/process/AgentService.ts with the full typed API (tag, Service, Session, helpers, types)
- Create @dxos/functions/src/process/RoutineError.ts (moved from @dxos/assistant)
- Add AgentService subpath export in @dxos/functions package.json and moon.yml compile entrypoint
- @dxos/assistant re-exports the full API from @dxos/functions for backward compatibility and keeps the layer/makeSession impl
- Update tsconfig project references to reflect the new dep graph (blueprints no longer references functions; functions now references blueprints)
