+++ packages/core/assistant/tsconfig.json
@@ -62,6 +62,9 @@
       "path": "../echo/echo-protocol"
     },
     {
+      "path": "../functions"
+    },
+    {
       "path": "../functions-runtime"
     },
     {
