+++ packages/core/blueprints/package.json
@@ -27,7 +27,6 @@
     "@dxos/ai": "workspace:*",
     "@dxos/echo": "workspace:*",
     "@dxos/errors": "workspace:*",
-    "@dxos/functions": "workspace:*",
     "@dxos/invariant": "workspace:*",
     "@dxos/keys": "workspace:*",
     "@dxos/log": "workspace:*",
