+++ packages/core/functions/moon.yml
@@ -8,5 +8,6 @@ tasks:
   compile:
     args:
       - '--entryPoint=src/index.ts'
+      - '--entryPoint=src/process/AgentService.ts'
       - '--entryPoint=src/Trace.ts'
       - '--platform=neutral'
