+++ packages/core/functions/package.json
@@ -18,6 +18,11 @@
       "types": "./dist/types/src/index.d.ts",
       "default": "./dist/lib/neutral/index.mjs"
     },
+    "./AgentService": {
+      "source": "./src/process/AgentService.ts",
+      "types": "./dist/types/src/process/AgentService.d.ts",
+      "default": "./dist/lib/neutral/process/AgentService.mjs"
+    },
     "./Trace": {
       "source": "./src/Trace.ts",
       "types": "./dist/types/src/Trace.d.ts",
@@ -33,6 +38,7 @@
   "scripts": {},
   "dependencies": {
     "@dxos/ai": "workspace:*",
+    "@dxos/blueprints": "workspace:*",
     "@dxos/context": "workspace:*",
     "@dxos/echo": "workspace:*",
     "@dxos/echo-db": "workspace:*",
