+++ packages/core/functions/tsconfig.json
@@ -41,6 +41,9 @@
       "path": "../ai"
     },
     {
+      "path": "../blueprints"
+    },
+    {
       "path": "../echo/echo"
     },
     {
